refactor: make GetOwners generic
